LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/intercept-handler-returns-non-promise-expected.txt


FAIL event.intercept() should resolve immediately if the handler doesn't return a promise assert_equals: expected "#1" but got ""
